backend/requirements.txt
# ── Base ─────────────────────────────────────────────
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# ── Database ─────────────────────────────────────────
sqlalchemy>=2.0.25,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0
alembic>=1.13.0,<2.0.0

# ── Celery + Redis ───────────────────────────────────
celery[redis]>=5.3.6,<6.0.0
redis>=5.0.0,<6.0.0

# ── HTTP client ──────────────────────────────────────
httpx>=0.27.0,<1.0.0
tenacity>=8.2.0,<9.0.0

# ── Security & Rate limiting ────────────────────────
slowapi>=0.1.9,<1.0.0
python-multipart>=0.0.9,<1.0.0
python-magic>=0.4.27,<1.0.0

# ── Document processing ─────────────────────────────
PyPDF2>=3.0.0,<4.0.0
python-docx>=1.1.0,<2.0.0
tiktoken>=0.6.0,<1.0.0

# ── ML (training) ───────────────────────────────────
torch>=2.1.0
transformers>=4.38.0,<5.0.0
trl>=0.7.0,<1.0.0
peft>=0.8.0,<1.0.0
bitsandbytes>=0.42.0
accelerate>=1.0.0,<2.0.0
datasets>=2.17.0,<3.0.0
huggingface-hub>=0.20.0,<1.0.0

# ── Logging ──────────────────────────────────────────
structlog>=24.1.0,<25.0.0

# ── Utils ────────────────────────────────────────────
jinja2>=3.1.3,<4.0.0

# ── Testing ──────────────────────────────────────────
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
httpx  # already above, used for TestClient
